chore(connect): split verifyAuthenticityProof.test.ts
